requirements.txt
# Web Framework
flask==3.0.0
werkzeug==3.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Deep Learning (v3.0)
tensorflow==2.15.0
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# AI & API
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Optional: Performance & Production
gunicorn==21.2.0
waitress==2.1.2

# Optional: Data Validation
jsonschema==4.20.0
